fix(wextra): fix proactive -Wsign-compare in test files (round 7)
- test_matmul_{multi_core_single_dram,large_block,single_core,single_core_small}.cpp: int vs data.size()
- test_dram_to_l1_multicast*.cpp: int vs grid_size.y/x (size_t)
- routing_test_common.hpp: int vs phys_cores.size()
- conv_pattern.hpp: int vs uint32_t row/col tile sizes

+++ tests/tt_metal/tt_metal/test_matmul_large_block.cpp
@@ -25,7 +25,7 @@ void print_faces(std::vector<bfloat16> data, string name) {
 
     int tile_index = 0;
     int face_index = 0;
-    for (int i = 0; i < data.size(); i++) {
+    for (size_t i = 0; i < data.size(); i++) {
         if (i % 256 == 0) {
             std::cout << "Tile " << tile_index / 4 << std::endl;
             std::cout << "Face = " << face_index << std::endl;
